Split native hit target audit tests
