+++ docs/CODE_QUALITY_FILE_GRADES.md
@@ -19,7 +19,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | `source:quill-code-desktop` | 36 | 5280 | 97 | A+ | `DesktopAutomationNotifier.swift` A-<br>`QuillCodeDesktopWorkspaceActionCoordinator.swift` A<br>`QuillCodeDesktopTerminalCoordinator.swift` A |
 | `source:quillcode-linux-computer-use-smoke` | 1 | 114 | 97 | A+ | `main.swift` A+ |
 | `test:QuillCodeAgentTests` | 21 | 3839 | 98 | A+ | `AgentFinalAnswerBuilderTests.swift` A-<br>`AgentToolLoopTests.swift` A-<br>`TrustedRouterActionParserTests.swift` A |
-| `test:QuillCodeAppTests` | 206 | 31428 | 98 | A+ | `QuillCodeNativeHitTargetAuditTests.swift` A-<br>`WorkspaceRemoteProjectToolExecutorTests.swift` A-<br>`WorkspaceBrowserIntegrationTests.swift` A- |
+| `test:QuillCodeAppTests` | 210 | 31466 | 98 | A+ | `WorkspaceRemoteProjectToolExecutorTests.swift` A-<br>`WorkspaceBrowserIntegrationTests.swift` A-<br>`WorkspaceReviewIntegrationTests.swift` A- |
 | `test:QuillCodeCoreTests` | 3 | 524 | 98 | A+ | `CoreModelTests.swift` A+<br>`VerificationResultParserTests.swift` A+<br>`ModelTokenUsageTests.swift` A+ |
 | `test:QuillCodeDesktopTests` | 3 | 1108 | 97 | A+ | `QuillCodeDesktopControllerSmokeTests.swift` A<br>`QuillCodeApprovalNotificationTests.swift` A+<br>`QuillCodeDesktopRenderedSmokeTests.swift` A+ |
 | `test:QuillCodeParityTests` | 145 | 13473 | 99 | A+ | `ParitySwiftInteractionTargetAuditSupport.swift` A-<br>`ParityWorkspaceReviewSurfaceGateTests.swift` A-<br>`ParityMCPGateTests.swift` A- |
@@ -32,7 +32,6 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 
 | Grade | Score | Lines | Module | File | Main Issues |
 | --- | ---: | ---: | --- | --- | --- |
-| A- | 89 | 969 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeNativeHitTargetAuditTests.swift` | large file (969 lines); 18 lines >120 chars; 2 lines >160 chars |
 | A- | 89 | 615 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/WorkspaceRemoteProjectToolExecutorTests.swift` | 5 lines >120 chars; 2 lines >160 chars; high duplicate-line ratio (33%) |
 | A- | 89 | 598 | `test:QuillCodeParityTests` | `Tests/QuillCodeParityTests/ParitySwiftInteractionTargetAuditSupport.swift` | 9 lines >120 chars; 1 lines >160 chars; high duplicate-line ratio (33%) |
 | A- | 89 | 423 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/WorkspaceBrowserIntegrationTests.swift` | 6 lines >120 chars; 2 lines >160 chars; high duplicate-line ratio (36%) |
@@ -72,6 +71,7 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A- | 91 | 325 | `test:QuillCodeAgentTests` | `Tests/QuillCodeAgentTests/AgentToolLoopTests.swift` | 2 lines >120 chars; high duplicate-line ratio (53%) |
 | A- | 91 | 323 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/WorkspaceAutomationRunIntegrationTests.swift` | 5 lines >120 chars; high duplicate-line ratio (33%) |
 | A- | 91 | 317 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/WorkspaceReviewSurfaceBuilderTests.swift` | 1 lines >120 chars; high duplicate-line ratio (25%) |
+| A- | 91 | 315 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeTopBarSurfaceTests.swift` | 3 lines >120 chars; high duplicate-line ratio (24%) |
 
 ## Every File
 
@@ -783,7 +783,11 @@ Generated by `scripts/grade-code-quality.py`. Grades are deterministic heuristic
 | A+ | 100 | 138 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeBrowserSurfaceTests.swift` | no automated issues |
 | A+ | 100 | 80 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeCommandIconCatalogTests.swift` | no automated issues |
 | A+ | 99 | 194 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeMCPSupportTests.swift` | 2 lines >120 chars |
-| A- | 89 | 969 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeNativeHitTargetAuditTests.swift` | large file (969 lines); 18 lines >120 chars; 2 lines >160 chars |
+| A+ | 98 | 443 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeNativeHitTargetAuditTestSupport.swift` | 4 lines >120 chars; 1 lines >160 chars |
+| A+ | 99 | 95 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeNativeHitTargetDesignTests.swift` | 2 lines >120 chars |
+| A+ | 99 | 114 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeNativeHitTargetSourceAuditTests.swift` | 1 lines >120 chars |
+| A+ | 99 | 210 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeNativeHitTargetSurfaceAuditTests.swift` | 4 lines >120 chars |
+| A+ | 97 | 145 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeNativeHitTargetValidationTests.swift` | 9 lines >120 chars; 1 lines >160 chars |
 | A+ | 100 | 46 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeProjectListSurfaceTests.swift` | no automated issues |
 | A+ | 96 | 405 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeReviewSurfaceTests.swift` | duplicate-line ratio (21%) |
 | A+ | 100 | 73 | `test:QuillCodeAppTests` | `Tests/QuillCodeAppTests/QuillCodeRuntimeIssueRecoveryPlannerTests.swift` | no automated issues |
